==> 1_setup_env.ipynb <==
!pip install kagglehub

import kagglehub
import pandas as pd
from sklearn.model_selection import train_test_split
import boto3
import os

# Download the dataset
path = kagglehub.dataset_download("mlg-ulb/creditcardfraud")
print("Path to dataset files:", path)

# Load the data
data = pd.read_csv(f'{path}/creditcard.csv')

# Data redaction (if needed)
# Example: data.drop(columns=['sensitive_column'], inplace=True)

# Data preprocessing
data = data.dropna()  # Remove missing values
X = data.drop(columns=['Class'])
y = data['Class']

# Split the data
X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.2, random_state=42)

# Create directories if they don't exist
os.makedirs('train', exist_ok=True)
os.makedirs('val', exist_ok=True)

# Save preprocessed data to S3
X_train.to_csv('train/X_train.csv', index=False)
y_train.to_csv('train/y_train.csv', index=False)
X_val.to_csv('val/X_val.csv', index=False)
y_val.to_csv('val/y_val.csv', index=False)

s3 = boto3.client('s3')
bucket_name = 'fr-detector'
s3.upload_file('train/X_train.csv', bucket_name, 'train/X_train.csv')
s3.upload_file('train/y_train.csv', bucket_name, 'train/y_train.csv')
s3.upload_file('val/X_val.csv', bucket_name, 'val/X_val.csv')
s3.upload_file('val/y_val.csv', bucket_name, 'val/y_val.csv')
